Move reducers stuff to /ts/store/reducers (#211)
