+++ ts/boot/configureStoreAndPersistor.ts
@@ -13,10 +13,10 @@ import createSagaMiddleware from "redux-saga";
 import thunk from "redux-thunk";
 
 import { analytics } from "../middlewares";
-import rootReducer from "../reducers";
-import { GlobalState } from "../reducers/types";
 import rootSaga from "../sagas";
 import { Action, Store, StoreEnhancer } from "../store/actions/types";
+import rootReducer from "../store/reducers";
+import { GlobalState } from "../store/reducers/types";
 import { NAVIGATION_MIDDLEWARE_LISTENERS_KEY } from "../utils/constants";
 
 const isDebuggingInChrome = __DEV__ && !!window.navigator.userAgent;
