+++ ts/screens/authentication/IdpLoginScreen.tsx
@@ -6,13 +6,13 @@ import { connect } from "react-redux";
 import AppHeader from "../../components/ui/AppHeader";
 import * as config from "../../config";
 import I18n from "../../i18n";
-import { GlobalState } from "../../reducers/types";
 import { loginFailure, loginSuccess } from "../../store/actions/session";
 import { ReduxProps } from "../../store/actions/types";
 import {
   isUnauthenticatedWithoutIdpSessionState,
   SessionState
 } from "../../store/reducers/session";
+import { GlobalState } from "../../store/reducers/types";
 import { extractLoginResult } from "../../utils/login";
 type ReduxMappedProps = {
   session: SessionState;
