+++ ts/screens/authentication/SpidInformationRequestScreen.tsx
@@ -19,7 +19,7 @@ import SpidInformationForm, {
 import AppHeader from "../../components/ui/AppHeader";
 import Modal from "../../components/ui/Modal";
 import I18n from "../../i18n";
-import { GlobalState } from "../../reducers/types";
+import { GlobalState } from "../../store/reducers/types";
 type ReduxMappedProps = {
   isFormValid: boolean;
 };
