+++ ts/store/reducers/index.ts
@@ -6,16 +6,16 @@ import { reducer as networkReducer } from "react-native-offline";
 import { Reducer, ReducersMapObject } from "redux";
 import { FormStateMap, reducer as formReducer } from "redux-form";
 
-import { Action } from "../store/actions/types";
-import entitiesReducer from "../store/reducers/entities";
-import errorReducer from "../store/reducers/error";
-import loadingReducer from "../store/reducers/loading";
-import notificationsReducer from "../store/reducers/notifications";
-import onboardingReducer from "../store/reducers/onboarding";
-import profileReducer from "../store/reducers/profile";
-import sessionReducer from "../store/reducers/session";
-import appStateReducer from "./appState";
+import appStateReducer from "../../sore/reducers/appState";
+import { Action } from "../actions/types";
+import entitiesReducer from "./entities";
+import errorReducer from "./error";
+import loadingReducer from "./loading";
 import navigationReducer from "./navigation";
+import notificationsReducer from "./notifications";
+import onboardingReducer from "./onboarding";
+import profileReducer from "./profile";
+import sessionReducer from "./session";
 import { GlobalState } from "./types";
 
 /**
