+++ ts/store/reducers/types.ts
@@ -1,15 +1,15 @@
 import { NavigationState } from "react-navigation";
 import { FormStateMap } from "redux-form";
 
-import { Action } from "../store/actions/types";
-import { EntitiesState } from "../store/reducers/entities";
-import { ErrorState } from "../store/reducers/error";
-import { LoadingState } from "../store/reducers/loading";
-import { NotificationsState } from "../store/reducers/notifications";
-import { OnboardingState } from "../store/reducers/onboarding";
-import { ProfileState } from "../store/reducers/profile";
-import { SessionState } from "../store/reducers/session";
-import { AppState } from "./appState";
+import { AppState } from "../../sore/reducers/appState";
+import { Action } from "../actions/types";
+import { EntitiesState } from "./entities";
+import { ErrorState } from "./error";
+import { LoadingState } from "./loading";
+import { NotificationsState } from "./notifications";
+import { OnboardingState } from "./onboarding";
+import { ProfileState } from "./profile";
+import { SessionState } from "./session";
 
 export type NetworkState = Readonly<{
   isConnected: boolean;
